requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
deprecation==2.1.0
distro==1.9.0
dnspython==2.4.2
docopt==0.6.2
email-validator==2.1.0.post1
fastapi==0.108.0
flower==2.0.1
frozenlist==1.4.1
gotrue==2.1.0
h11==0.14.0
httpcore==0.17.3
httptools==0.6.1
httpx==0.24.1
humanize==4.9.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.4
MarkupSafe==2.1.3
multidict==6.0.4
openai==1.6.1
orjson==3.9.10
packaging==23.2
pipreqs==0.4.13
postgrest==0.13.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-extra-types==2.3.0
pydantic-settings==2.1.0
pydantic_core==2.14.6
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
realtime==1.0.2
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.32.0.post1
storage3==0.7.0
StrEnum==0.4.15
supabase==2.3.0
supafunc==0.3.1
tornado==6.4
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
ujson==5.9.0
urllib3==2.1.0
uvicorn==0.25.0
uvloop==0.19.0
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.12
websockets==11.0.3
yarg==0.1.9
yarl==1.9.4
